tests/requirements.txt
# Testing framework
pytest>=7.4.0
pytest-html>=3.2.0
pytest-json-report>=1.5.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0

# HTTP client for testing MCP servers
requests>=2.31.0
httpx>=0.24.0

# JSON schema validation
jsonschema>=4.19.0

# Test utilities
faker>=19.3.0
parameterized>=0.9.0

# Retry logic for flaky tests
tenacity>=8.2.0

# Database clients for direct testing
psycopg2-binary>=2.9.7
pymysql>=1.1.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0